+++ stack/core/variables.tf
@@ -13,4 +13,12 @@ variable "ecr_lifecycle_policy_rules" {
   type    = list(any)
   default = []
 }
+
+variable "project" {
+  type = string
+}
+
+variable "environment" {
+  type = string
+}
 # =============== ECR ================ #