+++ stack/global/deploy-bot.tf
@@ -1,0 +1,3 @@
+module "deploy-bot" {
+  source = "../../iam//iam-for-deploy-bot"
+}